@@ -26,14 +26,14 @@ version = "2.0.0"
 packages = [{ include = "coral" }]
 
 [tool.poetry.dependencies]
-cvxopt = "*"
-cnvkit = "*"
+cvxopt = ">=1.3.2"
+cnvkit = { git = "https://github.com/etal/cnvkit.git" }
 intervaltree = "*"
 matplotlib = "*"
-pandas = { version = "*", python = ">=3.7,<4.0" }
+pandas = { version = ">=2.2.3", python = "=3.12.*" }
 pysam = { version = ">0.1.7" }
 pytest = { version = ">=4.4", optional = true }
-python = ">=3.10,<4.0"
+python = "=3.12.*"
 numpy = "==2.0.2"
 numba = "==0.60.0"
 llvmlite = "==0.43.0"
@@ -97,7 +97,7 @@ lint.ignore = [
   "RUF005",  # Prefer iterable unpacking over concatenation
   "PLR0911", # too-many-return-statements
   "FIX002",  # Line contains TODO
-  
+  "TRY400",  # logger.error vs. logger.exception (don't want to log traceback)
 ]
 lint.extend-select = ["I"]
 line-length = 80
@@ -107,8 +107,10 @@ lint.pycodestyle.max-doc-length = 80
 enable_error_code = "possibly-undefined"
 disable_error_code = "import-untyped"
 exclude = [
-  "hsr.py", "scripts/report_nanopore_qc.py",
-  "plot_amplicons.py", "plot_cn.py"
+  "hsr.py", 
+  "scripts/report_nanopore_qc.py",
+  "plot_amplicons.py", 
+  "plot_cn.py"
 ]
 plugins = "numpy.typing.mypy_plugin"
 strict = false
